pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hedera-py-lite"
version = "0.1.0"
description = "Lightweight, JVM-free Python client for the Hedera network"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Emmanuel Okechukwu Nwajari", email = "De-real-iManuel@hotmail.com" },
]
keywords = ["hedera", "hashgraph", "hbar", "hcs", "blockchain", "grpc", "web3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
    "Typing :: Typed",
]
dependencies = [
    "grpcio>=1.60.0",
    "cryptography>=42.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/De-real-iManuel/hedera-py-lite"
Repository = "https://github.com/De-real-iManuel/hedera-py-lite"
"Bug Tracker" = "https://github.com/De-real-iManuel/hedera-py-lite/issues"
Changelog = "https://github.com/De-real-iManuel/hedera-py-lite/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
